move HMC-related functions to inference.hmc
