--- blackjax/inference/__init__.py
@@ -1,3 +1,0 @@
-from . import integrators
-
-__all__ = ["integrators"]
