+++ blackjax/inference/hmc/__init__.py
@@ -1,0 +1,3 @@
+from . import integrators
+
+__all__ = ["integrators"]
