+++ blackjax/inference/hmc/metrics.py
