+++ blackjax/inference/hmc/termination.py
